chore: Refactor builder to be more useful for NCL
As implemented the builder always expected a JSON string to be passed
containing the parsed directives. This is likely for the CLI usecase but
involves an unnecessary and very expensive step of
marshalling/unmarshalling if you already have the structure in memory.

These changes allow you to build a config from a slice of Directives and
moves something similar to the existing function to a `BuildFromJSON`
function.

In addition the implementation of "BuildBlock" has been refactored to
use `strings.Builder`-- which is much more efficient when building
large strings-- and to make the function unexported.

The following benchmarks demonstrate improvements over the old
implementation and the new (BuildX).

```
BenchmarkBuild-8    	    3442	    341209 ns/op	  250544 B/op	    1388 allocs/op
BenchmarkBuildX-8   	   32696	     36263 ns/op	   50560 B/op	     583 allocs/op
```