Adds test coverage to Lexer and updates string behavior
Found potentially unintended difference between port and the reference
implementation where quoted text included the quotes in the token.

Quotes are a funny thing in an NGINX configuration where they aren't
usually necessary unless they are disambiguating parts of the
configuration. Other crossplane implementations track if the characters
are "quoted" as a hint as to how to treat the rest of the characters but
not for much else. If you read a config and print it out then the usage
of quotes will appear differently if the quotes are unnecessary.