feat: Updates ParseError to export fields
`ParseError` had no exported fields which made it a bit of a useless
part of crossplane's API. This is especially annoying when it's gets
passed to the callback function and there's nothing you can really do
but acknowledge that there was an error.

While I was in there I noticed that the `ParseError` struct was being
passed around not as a pointer but as a copy of the struct. Since it's
just a container for a couple pointers and a string this isn't really
a big deal but it made it a little odd any annoying to use with
`errors.As`, so I changed it.

Not touching the `MarshalJSON` method that returns just a string. That's
weird but it's also technically ok.

```node
> console.log(JSON.stringify("this is actually ok"))
"this is actually ok"
```