fix: validate body of types directive
types is a "map-like" directive meaning its body does not contain additional NGINX directives. Previously, if the body of the types directive was not empty, crossplane would fail to parse it because it did not recognize its contents as NGINX directives. This fixes this behavior by adding types to the set of map-like directives and validating only that each entry has the correct number of arguments, 1 or more.