Fixes error code on test run
`TestMain` is used for all tests, not just benchmarks as it looks like
the file might expect. The error code was being ignored which results in
an `ok` status on each test run even when there are failures.

This commit fixes `TestMain` to return the error code returned from
`Run()`. I did not look at anything else going on in the file.