Adds parse option for custom directives, moves nap directives
Added an API for allowing the caller of `Parse` to specify bitmasks for
directives outsides the "core" set of NGINX directives. This would allow
the caller to parse an NGINX configuration that contained directives
from a dynamic module even if it is a unknown custom one as long as the
configuration still matches the expected grammar.

As part of this I moved the NAP WAF v4 directives out of the main
`directives` map and added directives for WAF v5.